--- govcd/generic_functions_test.go
@@ -1,96 +1,0 @@
-package govcd
-
-import (
-	"reflect"
-	"testing"
-)
-
-func Test_oneOrError(t *testing.T) {
-	type args struct {
-		key         string
-		name        string
-		entitySlice []*testEntity
-	}
-	tests := []struct {
-		name                  string
-		args                  args
-		want                  *testEntity
-		wantErr               bool
-		wantErrEntityNotFound bool
-	}{
-		{
-			name: "SingleEntity",
-			args: args{
-				key:         "name",
-				name:        "test",
-				entitySlice: []*testEntity{{name: "test"}},
-			},
-			want:    &testEntity{name: "test"},
-			wantErr: false,
-		},
-		{
-			name: "NoEntities",
-			args: args{
-				key:         "name",
-				name:        "test",
-				entitySlice: []*testEntity{},
-			},
-			want:                  nil,
-			wantErr:               true,
-			wantErrEntityNotFound: true,
-		},
-		{
-			name: "TwoEntities",
-			args: args{
-				key:         "name",
-				name:        "test",
-				entitySlice: []*testEntity{{name: "test"}, {name: "best"}},
-			},
-			want:    nil,
-			wantErr: true,
-		},
-		{
-			name: "ThreeEntities",
-			args: args{
-				key:         "name",
-				name:        "test",
-				entitySlice: []*testEntity{{name: "test"}, {name: "best"}, {name: "rest"}},
-			},
-			want:    nil,
-			wantErr: true,
-		},
-		{
-			name: "NilEntities",
-			args: args{
-				key:         "name",
-				name:        "test",
-				entitySlice: nil,
-			},
-			want:                  nil,
-			wantErr:               true,
-			wantErrEntityNotFound: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := oneOrError(tt.args.key, tt.args.name, tt.args.entitySlice)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("oneOrError() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-
-			if tt.wantErr && tt.wantErrEntityNotFound && !ContainsNotFound(err) {
-				t.Errorf("oneOrError() error = %v, wantErrEntityNotFound %v", err, tt.wantErrEntityNotFound)
-				return
-			}
-
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("oneOrError() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-type testEntity struct {
-	name string
-}
